requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.1
altair==5.4.1
annotated-types==0.7.0
anyio==4.7.0
asyncio==3.4.3
async_generator==1.10
attrs==24.2.0
backoff==2.2.1
beautifulsoup4==4.12.3
blinker==1.8.2
bs4==0.0.2
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
ConfigArgParse==1.7
contourpy==1.3.0
cycler==0.12.1
diff-match-patch==20241021
diskcache==5.6.3
distro==1.9.0
entrypoints==0.4
Faker==30.8.0
favicon==0.7.0
filelock==3.16.1
flake8==7.1.1
fonttools==4.54.1
frozenlist==1.5.0
fsspec==2024.10.0
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.0.3
grep-ast==0.4.1
h11==0.14.0
htbuilder==0.6.2
httpcore==1.0.7
httpx==0.27.2
huggingface-hub==0.26.5
idna==3.10
importlib_metadata==7.2.1
importlib_resources==6.4.5
Jinja2==3.1.4
jiter==0.8.0
json5==0.10.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.7
llvmlite==0.43.0
litellm==1.55.3
lxml==5.3.0
Markdown==3.7
markdown-it-py==3.0.0
markdownlit==0.0.7
MarkupSafe==3.0.2
matplotlib==3.9.2
mccabe==0.7.0
mdurl==0.1.2
mixpanel==4.10.1
monotonic==1.6
more-itertools==10.5.0
mpmath==1.3.0
multidict==6.1.0
narwhals==1.10.0
networkx==3.2.1
numba==0.60.0
numpy==1.26.4
ollama==0.4.4
openai==1.57.0
openai-whisper==20240930
packaging==24.2
pandas==2.2.3
pathspec==0.12.1
pexpect==4.9.0
pillow==10.4.0
playwright==1.47.0
plotly==5.24.1
posthog==3.7.4
prometheus_client==0.21.0
prompt_toolkit==3.0.48
propcache==0.2.1
protobuf==5.28.2
psutil==6.1.0
ptyprocess==0.7.0
pyarrow==17.0.0
PyAudio==0.2.14
pycodestyle==2.12.1
pycparser==2.22
pydantic==2.10.3
pydantic_core==2.27.1
pydeck==0.9.1
pydub==0.25.1
pyee==12.0.0
pyflakes==3.2.0
Pygments==2.18.0
pymdown-extensions==10.11.2
pypandoc==1.14
pyparsing==3.2.0
pyperclip==1.9.0
pytest>=8.3.4 
pytest-asyncio>=0.25.0
python-ffmpeg>=2.0.12
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
scipy==1.13.1
setuptools==75.1.0
six==1.17.0
smmap==5.0.1
sniffio==1.3.1
sounddevice==0.5.1
soundfile==0.12.1
soupsieve==2.6
st-annotated-text==4.0.1
st-theme==1.2.3
streamlit==1.39.0
streamlit-autorefresh==1.0.1
streamlit-camera-input-live==0.2.0
streamlit-card==1.0.2
streamlit-embedcode==0.1.2
streamlit-extras==0.5.0
streamlit-faker==0.0.3
streamlit-image-coordinates==0.1.9
streamlit-keyup==0.2.4
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==2.5.5
sympy==1.13.1
tenacity==9.0.0
tiktoken==0.8.0
tokenizers==0.19.1
toml==0.10.2
torch==2.5.0
tornado==6.4.1
tqdm==4.67.1
tree-sitter==0.21.3
tree-sitter-languages==1.10.2
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
validators==0.34.0
watchdog==5.0.3
watchfiles==1.0.0
wcwidth==0.2.13
websockets==14.1
wheel==0.44.0
yarl==1.18.3
zipp==3.21.0
